+++ openfisca_aotearoa/tests/social_security/accommodation_supplement/2018/costs.yaml
@@ -1,7 +1,7 @@
 - name: >
     Social Security Act 2018 -
-    65 Accommodation Supplement
-    (2) Accommodation Costs
+    65 Accommodation Supplement -
+    (2) Accommodation Costs -
     (a) Rent
   period: 2018-11-26
   input:
