+++ openfisca_aotearoa/tests/social_security/accommodation_supplement/2018/situation.yaml
@@ -1,6 +1,6 @@
 - name: >
     Social Security Act 2018 -
-    Acommodation Supplement -
+    Accommodation Supplement -
     Situation -
     Person who has 1 or more dependent children and who is in a relationship
   period: 2018-11-26
@@ -186,7 +186,7 @@
           2022-W26: 4
           2022-W27: 0
           2022-W28: 0
-          
+
 
 - name: >
     Social Security Act 2018 -
